done api requests for cleaners list get, cleaner get/add/update/delete

+++ myapi/models.py
@@ -1,7 +1,8 @@
 from django.db import models
 
 class CleanersInfo(models.Model):
-    id = models.IntegerField(primary_key=True, null=False, unique=True, auto_created=True)
+    # id = models.IntegerField(primary_key=True, null=False, unique=True, auto_created=True)
+    id = models.BigAutoField(primary_key=True, null=False, unique=True, auto_created=True)
     name = models.CharField(max_length=60, blank=False)
     surname = models.CharField(max_length=60, blank=False)
     city = models.CharField(max_length=60, blank=True)
